+++ src/components/Main.vue
@@ -35,8 +35,10 @@
             </div>
             <SiteTypeList></SiteTypeList>
             <AgencySelect></AgencySelect>
+            <br />
             <WatershedInput></WatershedInput>
             <TemporalRange></TemporalRange>
+            <SiteStatusSelect></SiteStatusSelect>
           </div>
         </div>
         <div class="rightcolumn">
@@ -69,6 +71,7 @@ import SiteTypeList from "../components/SiteTypeList";
 import CoordinatesInput from "../components/CoordinatesInput";
 import HUCInput from "../components/HUCInput";
 import ParamSelect from "../components/ParamSelect";
+import SiteStatusSelect from "../components/SiteStatusSelect";
 import AgencySelect from "../components/AgencySelect";
 import { mapState } from "vuex";
 import { notify } from "../notifications.js";
@@ -92,6 +95,7 @@ export default {
     ParamSelect,
     CountySelect,
     AgencySelect,
+    SiteStatusSelect,
     ToolTip,
     WatershedInput,
     TemporalRange
@@ -151,6 +155,7 @@ export default {
         siteTypeList: this.$store.getters.siteType,
         agencyCodeActive: this.$store.getters.agencyActive,
         agencyCode: this.$store.getters.agencyCode,
+        siteStatus: this.$store.getters.siteStatus,
         watershedAreaBounds: this.$store.getters.watershedAreaBounds,
         watershedUpperAreaBoundsActive: this.$store.getters
           .watershedUpperAreaBoundsActive,