+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added options for Test and QA buckets as targets for Jenkins builds
 ### Changed
 - Changed styling so all elements are left justified
+- Added site status parameter
 - Added watershed area query
 - Added csv processing for quickly adding a lot codes in fields which accept and verify multiple codes
 - Added various temporal query parameters